+++ x/rewards/types/keys.go
@@ -86,3 +86,14 @@ var (
 	// Value: None
 	RewardsRecordAddressIndexPrefix = []byte{0x02}
 )
+
+// FlatFee prefixed store state keys.
+var (
+	// FlatFeeStatePrefix defines the state global prefix.
+	FlatFeeStatePrefix = []byte{0x05}
+
+	// FlatFeePrefix defines the prefix for storing FlatFee objects.
+	// Key: FlatFeeStatePrefix | FlatFeePrefix | {ContractAddress}
+	// Value: sdk.Coin
+	FlatFeePrefix = []byte{0x00}
+)