+++ CHANGELOG.md
@@ -31,6 +31,7 @@ Contains all the PRs that improved the code without changing the behaviours.
 ### Added
 
 - [#253](https://github.com/archway-network/archway/pull/253) - add wasm bindings for contracts to query the x/gov module.
+- [#254](https://github.com/archway-network/archway/pull/254) - implementing contract premiums (keeper)
 
 ## [v0.1.0]